Split `resource_endpoints_test.py` as it is getting too big
SDESK-7484
